separating into discrete components and improving types
