+++ src/client/Components/Page/Page.tsx
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import DataTable from '../DataTable/DataTable';
+import TableData from '../DataTable/TableData/TableData';
 
 const Page = () => {
+  // let jsonData: taskData;
   const [dataLoaded, setDataLoaded] = useState(null);
   const [jsonData, setJsonData] = useState(null);
 
@@ -10,6 +12,11 @@ const Page = () => {
   //   // const now: Date = new Date();
   //   // console.log(Date.getFullYear(Date.now()));
   // };
+
+  const updateClickHandler = (event: React.MouseEvent<HTMLButtonElement>): void => {
+    event.preventDefault();
+    console.log(event.currentTarget.name);
+  };
 
   useEffect(() => {
     fetch('http://localhost:8000/tasks', {
@@ -31,7 +38,9 @@ const Page = () => {
     return <h1>Loading...</h1>;
   }
   return (
-    <DataTable jsonData={jsonData} />
+    <DataTable>
+      <TableData jsonData={jsonData} updateClickHandler={updateClickHandler} />
+    </DataTable>
   );
 };
